feat: 장바구니 계산 로직을 orderSummaryHandlers.ts로 이동 및 관련 주석 업데이트
장바구니 계산 및 UI 업데이트 로직을 `orderSummaryHandlers.ts`로 분리하여 코드의 가독성과 유지보수성을 향상시켰습니다. `main.basic.ts`와 여러 컴포넌트에서 `handleCalculateCartStuff` 함수를 호출하도록 수정하여 장바구니 관련 기능을 명확하게 관리할 수 있게 되었습니다.
